requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
datasets==2.7.1
dill==0.3.6
fastapi==0.88.0
filelock==3.8.2
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langdetect==1.0.9
libclang==14.0.6
lightning-transformers==0.2.5
lightning-utilities==0.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.3
multiprocess==0.70.14
mypy-extensions==0.4.3
nltk==3.8
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pathspec==0.10.3
Pillow==9.3.0
pip==22.3.1
platformdirs==2.6.0
protobuf==3.19.6
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.2
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.8.4.post0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
sentencepiece==0.1.97
setuptools==65.5.1
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
starlette==0.22.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.0
torch-tb-profiler==0.4.0
torchaudio==0.13.0
torchmetrics==0.11.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
wheel==0.38.4
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.2
